src/requirements.txt
asgiref==3.11.0
bandit==1.9.2
certifi==2025.11.12
cffi==2.0.0
cfgv==3.5.0
charset-normalizer==3.4.4
coverage==7.13.0
cryptography==46.0.3
defusedxml==0.7.1
distlib==0.4.0
Django==5.2.9
django-cors-headers==4.9.0
django-stubs==5.2.8
django-stubs-ext==5.2.8
djangorestframework==3.16.1
filelock==3.20.1
identify==2.6.15
idna==3.11
iniconfig==2.3.0
librt==0.7.4
markdown-it-py==4.0.0
mdurl==0.1.2
mwclient==0.11.0
mwoauth==0.4.0
mwparserfromhell==0.7.2
mypy==1.19.1
mypy_extensions==1.1.0
nodeenv==1.9.1
oauthlib==3.3.1
packaging==25.0
pathspec==0.12.1
platformdirs==4.5.1
pluggy==1.6.0
pre_commit==4.5.1
pycparser==2.23
Pygments==2.19.2
PyJWT==2.10.1
pytest==9.0.2
pytest-cov==7.0.0
pytest-django==4.11.1
python-social-auth==0.3.6
python3-openid==3.2.0
pywikibot==10.7.3
PyYAML==6.0.3
requests==2.32.5
requests-oauthlib==2.0.0
rich==14.2.0
ruff==0.14.9
social-auth-app-django==5.6.0
social-auth-core==4.8.1
sqlparse==0.5.4
stevedore==5.6.0
types-PyYAML==6.0.12.20250915
types-requests==2.32.4.20250913
typing_extensions==4.15.0
urllib3==2.6.1
virtualenv==20.35.4
